backend/requirements.txt
asgiref==3.4.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.2.0
django-model-utils==4.2.0
django-phone-field==1.8.1
django-polymorphic==3.1.0
django-rest-knox==4.1.0
django-rest-polymorphic==0.1.9
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==9.8.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.4
python-environ==0.4.54
pytz==2021.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.10